# api/openapi/catalog-v1.yaml
openapi: 3.0.3
info:
  title: Catalog REST API
  description: REST API for Kubeflow Hub Catalog
  version: v1
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url: "https://localhost:8080"
  - url: "http://localhost:8080"
paths:
...
  /api/model_catalog/v1/labels:
    summary: Path used to get the list of catalog labels.
    description: >-
      The REST endpoint/path used to list zero or more `CatalogLabel` entities.
    get:
      summary: List All CatalogLabels
      tags:
        - ModelCatalogService
      parameters:
        - $ref: "#/components/parameters/assetType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/labelOrderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogLabelListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findLabels
      description: Gets a list of all `CatalogLabel` entities.
  /api/model_catalog/v1/models:
    description: >-
      The REST endpoint/path used to list zero or more `CatalogModel` entities from all `CatalogSources`.
    get:
      summary: Search catalog models across sources.
      tags:
        - ModelCatalogService
      parameters:
        - name: recommendations
          in: query
          deprecated: true
          description: >-
            Deprecated: use `orderBy=RECOMMENDED` instead. Sort models by lowest recommended latency using Pareto filtering.
          required: false
          schema:
            type: boolean
            default: false
        - name: targetRPS
          in: query
          description: Target requests per second for latency calculations
          required: false
          schema:
            type: integer
            format: int32
        - name: latencyProperty
          in: query
          description: Property name for latency metric
          required: false
          schema:
            type: string
            default: "ttft_p90"
        - name: rpsProperty
          in: query
          description: Property name for RPS metric
          required: false
          schema:
            type: string
            default: "requests_per_second"
        - name: hardwareCountProperty
          in: query
          description: Property name for hardware count
          required: false
          schema:
            type: string
            default: "hardware_count"
        - name: hardwareTypeProperty
          in: query
          description: Property name for hardware type grouping
          required: false
          schema:
            type: string
            default: "hardware_type"
        - name: source
          description: |-
            Filter models by source. Multiple values can be separated by commas
            to filter by multiple sources (OR logic). For example:
            ?source=huggingface,local will return models from either
            huggingface OR local sources.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          in: query
          required: false
        - name: q
          description: Free-form keyword search used to filter the response.
          schema:
            type: string
          in: query
          required: false
        - name: sourceLabel
          description: |-
            Filter models by the label associated with the source. Multiple
            values can be separated by commas. If one of the values is the
            string `null`, then models from every source without a label will
            be returned.
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
        - $ref: "#/components/parameters/filterQuery"
        - $ref: "#/components/parameters/pageSize"
        - name: orderBy
          style: form
          explode: true
          examples:
            orderBy:
              value: ID
          description: |-
            Specifies the order by criteria for listing entities.

            Supported values are:
            - CREATE_TIME
            - LAST_UPDATE_TIME
            - ID
            - NAME
            - ACCURACY
            - RECOMMENDED

            Defaults to `NAME`.

            The `ACCURACY` sort will sort by the `overall_average` property in any linked metrics artifact.

            The `RECOMMENDED` sort applies Pareto filtering and ranks models by recommended latency.
            The Pareto-related parameters (`targetRPS`, `latencyProperty`, `rpsProperty`,
            `hardwareCountProperty`, `hardwareTypeProperty`) are applied the same way as with the
            deprecated `recommendations` parameter. With the default `sortOrder=ASC`, the most
            recommended models (lowest latency) appear first. Use `sortOrder=DESC` to show the least
            recommended configurations first.

            In addition, models can be sorted by properties. For example:
            - `provider.string_value` sorts by provider name
            - `artifacts.ifeval.double_value` sorts by the min/max value a property called ifeval across all associated artifacts
          schema:
            $ref: "#/components/schemas/OrderByField"
          in: query
          required: false
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogModelListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findModels
  /api/model_catalog/v1/models/filter_options:
    description: Lists options for `filterQuery` when listing models.
    get:
      summary: Lists fields and available options that can be used in `filterQuery` on the list models endpoint.
      tags:
        - ModelCatalogService
      responses:
        "200":
          $ref: "#/components/responses/FilterOptionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findModelsFilterOptions
  /api/model_catalog/v1/sources:
    summary: Path used to get the list of catalog sources.
    description: >-
      The REST endpoint/path used to list zero or more `CatalogSource` entities.
    get:
      summary: List All CatalogSources
      tags:
        - ModelCatalogService
      parameters:
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/assetType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/orderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogSourceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findSources
      description: Gets a list of all `CatalogSource` entities.
...
  /api/model_catalog/v1/sources/{source_id}/models/{model_name+}:
    description: >-
      The REST endpoint/path used to get a `CatalogModel`.
    get:
      summary: Get a `CatalogModel`.
      tags:
        - ModelCatalogService
      responses:
        "200":
          $ref: "#/components/responses/CatalogModelResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: getModel
    parameters:
      - name: source_id
        description: A unique identifier for a `CatalogSource`.
        schema:
          type: string
        in: path
        required: true
      - name: model_name+
        description: A unique identifier for the model.
        schema:
          type: string
        in: path
        required: true
  /api/model_catalog/v1/sources/{source_id}/models/{model_name}/artifacts:
    description: >-
      The REST endpoint/path used to list `CatalogArtifacts`.
    get:
      summary: List CatalogArtifacts.
      tags:
        - ModelCatalogService
      responses:
        "200":
          $ref: "#/components/responses/CatalogArtifactListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: getAllModelArtifacts
    parameters:
      - name: source_id
        description: A unique identifier for a `CatalogSource`.
        schema:
          type: string
        in: path
        required: true
      - name: model_name
        description: A unique identifier for the model.
        schema:
          type: string
        in: path
        required: true
      - $ref: "#/components/parameters/artifactType"
      - $ref: "#/components/parameters/artifact_type"
      - $ref: "#/components/parameters/artifactFilterQuery"
      - $ref: "#/components/parameters/pageSize"
      - $ref: "#/components/parameters/artifactOrderBy"
      - $ref: "#/components/parameters/sortOrder"
      - $ref: "#/components/parameters/nextPageToken"
  /api/model_catalog/v1/sources/{source_id}/models/{model_name}/artifacts/performance:
    description: >-
      Lists performance metrics artifacts (that is, artifacts of type `CatalogMetricsArtifact` where `metricsType` is `performance-metrics`).
    get:
      summary: List CatalogArtifacts.
      tags:
        - ModelCatalogService
      responses:
        "200":
          $ref: "#/components/responses/CatalogArtifactListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: getAllModelPerformanceArtifacts
    parameters:
      - name: source_id
        description: A unique identifier for a `CatalogSource`.
        schema:
          type: string
        in: path
        required: true
      - name: model_name
        description: A unique identifier for the model.
        schema:
          type: string
        in: path
        required: true
      - name: targetRPS
        description: |-
          Target requests per second. If specified, values for `replicas` and
          `total_requests_per_second` will be calculated and returned as custom
          properties.
        schema:
          type: integer
        in: query
      - name: recommendations
        description: Filter records that are less optimal based on approximate cost to run and latency.
        schema:
          type: boolean
          default: false
        in: query
      - name: rpsProperty
        description: Custom property name for requests per second metric.
        schema:
          type: string
          default: "requests_per_second"
        in: query
      - name: latencyProperty
        description: Custom property name for latency metric (e.g., ttft_p90, p90_latency).
        schema:
          type: string
          default: "ttft_p90"
        in: query
      - name: hardwareCountProperty
        description: Custom property name for hardware count metric.
        schema:
          type: string
          default: "hardware_count"
        in: query
      - name: hardwareTypeProperty
        description: Custom property name for hardware type grouping.
        schema:
          type: string
          default: "hardware_type"
        in: query
      - $ref: "#/components/parameters/artifactFilterQuery"
      - $ref: "#/components/parameters/pageSize"
      - $ref: "#/components/parameters/artifactOrderBy"
      - $ref: "#/components/parameters/sortOrder"
      - $ref: "#/components/parameters/nextPageToken"
...
components:
  schemas:
...
    CatalogAssetType:
      type: string
      enum:
        - models
        - mcp_servers
        - agents
        - skills
      default: models
...
  responses:
...
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: Bad Request parameters
...
    CatalogLabelListResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CatalogLabelList"
      description: A response containing a list of CatalogLabel entities.
...
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: Unauthorized
...
  parameters:
...
    labelOrderBy:
      style: form
      explode: true
      examples:
        labelOrderBy:
          value: name
      name: orderBy
      description: |
        Specifies the key to order catalog labels by. You can provide any string key
        that may exist in the label maps. Labels that contain the specified key will
        be sorted by that key's value. Labels that don't contain the key will maintain
        their original order and appear after labels that do contain the key.
      schema:
        type: string
      in: query
      required: false
    assetType:
      name: assetType
      description: Filter by asset type.
      schema:
        $ref: "#/components/schemas/CatalogAssetType"
      in: query
      required: false
...
    pageSize:
      examples:
        pageSize:
          value: 100
      name: pageSize
      description: Number of entities in each page.
      schema:
        type: string
        format: int32
        pattern: "^[1-9][0-9]{0,8}$"
      in: query
      required: false
    nextPageToken:
      name: nextPageToken
      description: >-
        Opaque pagination token returned by a previous list call. Do not construct manually; use the value from a prior response's nextPageToken field.
      schema:
        type: string
      in: query
      required: false
    sortOrder:
      style: form
      explode: true
      examples:
        sortOrder:
          value: DESC
      name: sortOrder
      description: "Specifies the sort order for listing entities, defaults to ASC."
      schema:
        $ref: "#/components/schemas/SortOrder"
      in: query
      required: false
...
  securitySchemes:
    Bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Bearer JWT scheme
security:
  - Bearer: []
...